# cmd/server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing proxy rules and iptables.",
        "title": "Go-Reauth-Proxy",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/license/MIT"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:7996",
    "basePath": "/",
    "paths": {
        "/api/auth": {
            "get": {
                "description": "Get the configured global authentication URLs and port",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "config"
                ],
                "summary": "Get global auth config",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.AuthConfig"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "Set the global authentication configurations (port, auth_url, login_url)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "config"
                ],
                "summary": "Set global auth config",
                "parameters": [
                    {
                        "description": "Auth configuration",
                        "name": "config",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AuthConfig"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/config/default-route": {
            "get": {
                "description": "Get the configured default route when root route is requested",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "config"
                ],
                "summary": "Get default route",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "Set the configured default route when root route is requested",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "config"
                ],
                "summary": "Set default route",
                "parameters": [
                    {
                        "description": "Route configuration, example: {\\",
                        "name": "rule",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/config/proxy-protocol": {
            "get": {
                "description": "Get whether the proxy port requires Proxy Protocol header",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "config"
                ],
                "summary": "Get proxy protocol force",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/admin.proxyProtocolForceResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "Enable or disable requiring Proxy Protocol header on proxy port",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "config"
                ],
                "summary": "Set proxy protocol force",
                "parameters": [
                    {
                        "description": "Proxy protocol options",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin.proxyProtocolForceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/admin.proxyProtocolForceResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/api/iptables/allow": {
            "post": {
                "description": "Add an ALLOW rule for a specific IP",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "iptables"
                ],
                "summary": "Allow IP",
                "parameters": [
                    {
                        "description": "IP to allow",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/iptables.ipRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/iptables/allow-all": {
            "post": {
                "description": "Remove the catch-all DROP rule",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "iptables"
                ],
                "summary": "Allow all traffic",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/iptables/block": {
            "post": {
                "description": "Add a DROP rule for a specific IP",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "iptables"
                ],
                "summary": "Block IP",
                "parameters": [
                    {
                        "description": "IP to block",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/iptables.ipRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/iptables/block-all": {
            "post": {
                "description": "Add a catch-all DROP rule at the end of the chain",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "iptables"
                ],
                "summary": "Block all traffic",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/iptables/clean": {
            "post": {
                "description": "Remove the custom iptables chain",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "iptables"
                ],
                "summary": "Clean iptables",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/iptables/flush": {
            "post": {
                "description": "Flush all rules in the custom chain",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "iptables"
                ],
                "summary": "Flush iptables rules",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/iptables/init": {
            "post": {
                "description": "Initialize the custom iptables chain",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "iptables"
                ],
                "summary": "Initialize iptables",
                "parameters": [
                    {
                        "description": "Initialization options",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/iptables.initRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "admin.proxyProtocolForceRequest": {
            "type": "object",
            "properties": {
                "proxy_protocol_force": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "admin.proxyProtocolForceResponse": {
            "type": "object",
            "properties": {
                "proxy_protocol_force": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "iptables.ipRequest": {
            "type": "object",
            "properties": {
                "ip": {
                    "type": "string",
                    "example": "192.168.1.100"
                }
            }
        },
        "models.AuthConfig": {
            "type": "object",
            "properties": {
                "auth_host": {
                    "type": "string",
                    "example": "auth.example.com"
                },
                "auth_port": {
                    "description": "Local Auth Service Port",
                    "type": "integer",
                    "example": 3000
                },
                "auth_url": {
                    "description": "Relative Verify URL (default /api/auth/verify)",
                    "type": "string",
                    "example": "/api/auth/verify"
                },
                "login_url": {
                    "description": "Relative Login URL (default /login)",
                    "type": "string",
                    "example": "/login"
                },
                "logout_url": {
                    "description": "Relative Logout URL (default /api/auth/logout)",
                    "type": "string",
                    "example": "/api/auth/logout"
                },
                "preflight_url": {
                    "description": "Relative Preflight URL (default /api/auth/preflight)",
                    "type": "string",
                    "example": "/api/auth/preflight"
                },
                "public_auth_base_url": {
                    "type": "string",
                    "example": "https://auth.example.com"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                },
                "timestamp": {
                    "type": "integer"
                }
            }
        }
    }
}